--- a/pkg/metrics_store/metrics_writer.go
+++ b/pkg/metrics_store/metrics_writer.go
@@ -92,12 +92,8 @@ func SanitizeHeaders(contentType string, writers MetricsWriterList) MetricsWrite
 	var lastHeader string
 	for _, writer := range writers {
 		if len(writer.stores) > 0 {
-			for i, header := range writer.stores[0].headers {
-
-				// Nothing to sanitize.
-				if len(header) == 0 {
-					continue
-				}
+			for i := 0; i < len(writer.stores[0].headers); {
+				header := writer.stores[0].headers[i]
 
 				// Removes duplicate headers from the given MetricsWriterList for the same family (generated through CRS).
 				// These are expected to be consecutive since G** resolution generates groups of similar metrics with same headers before moving onto the next G** spec in the CRS configuration.
@@ -106,8 +102,8 @@ func SanitizeHeaders(contentType string, writers MetricsWriterList) MetricsWrite
 
 					// If the requested content type was proto-based, replace "info" and "statesets" with "gauge", as they are not recognized by Prometheus' protobuf machinery.
 					if strings.HasPrefix(contentType, expfmt.ProtoType) {
-						infoTypeString := metric.Info.String()
-						stateSetTypeString := metric.StateSet.String()
+						infoTypeString := string(metric.Info)
+						stateSetTypeString := string(metric.StateSet)
 						if strings.HasSuffix(header, infoTypeString) {
 							header = header[:len(header)-len(infoTypeString)] + string(metric.Gauge)
 							writer.stores[0].headers[i] = header
@@ -121,22 +117,17 @@ func SanitizeHeaders(contentType string, writers MetricsWriterList) MetricsWrite
 
 				// Nullify duplicate headers after the sanitization to not miss out on any new candidates.
 				if header == lastHeader {
-					writer.stores[0].headers[i] = ""
+					writer.stores[0].headers = append(writer.stores[0].headers[:i], writer.stores[0].headers[i+1:]...)
+
+					// Do not increment the index, as the next header is now at the current index.
+					continue
 				}
 
 				// Update the last header.
 				lastHeader = header
-			}
-		}
-	}
 
-	// Remove all empty headers.
-	for _, writer := range writers {
-		if len(writer.stores) > 0 {
-			for i := len(writer.stores[0].headers) - 1; i >= 0; i-- {
-				if writer.stores[0].headers[i] == "" {
-					writer.stores[0].headers = append(writer.stores[0].headers[:i], writer.stores[0].headers[i+1:]...)
-				}
+				// Move to the next header.
+				i++
 			}
 		}
 	}