@rexagod 2023-12-15
A couple of notes around the latest patch.

* Manipulate headers only when they are not duplicates (since they will be removed before any operations).
* The earlier approach used `strings.Has*` and `strings.LastIndex` functions, with respective complexities of O(1) and O(n). This patch removes the dependency on the latter so the whole operation is lazily run in constant time.

### Benchmarks after the first optimization (comparing the first commit v/s second one on this PR)

```
goos: darwin
goarch: arm64
pkg: k8s.io/kube-state-metrics/v2/pkg/metrics_store
                                                  │   PUSHED    │               STAGED                │
                                                  │   sec/op    │   sec/op     vs base                │
SanitizeHeaders/text-format_unique_headers-10       561.4µ ± 0%   561.2µ ± 0%        ~ (p=0.811 n=10)
SanitizeHeaders/text-format_duplicate_headers-10    529.1µ ± 0%   281.6µ ± 0%  -46.77% (p=0.000 n=10)
SanitizeHeaders/proto-format_unique_headers-10      4.047m ± 0%   3.985m ± 0%   -1.54% (p=0.000 n=10)
SanitizeHeaders/proto-format_duplicate_headers-10   591.4µ ± 0%   279.1µ ± 0%  -52.81% (p=0.000 n=10)
geomean                                             918.2µ        647.5µ       -29.48%
```

### Benchmarks after the both optimizations (comparing the first commit v/s second one on this PR)

```
goos: darwin
goarch: arm64
pkg: k8s.io/kube-state-metrics/v2/pkg/metrics_store
                                                  │    PUSHED    │                 STAGED                 │
                                                  │    sec/op    │    sec/op      vs base                 │
SanitizeHeaders/text-format_unique_headers-10       561.39µ ± 0%    31.79µ ±  0%   -94.34% (p=0.000 n=10)
SanitizeHeaders/text-format_duplicate_headers-10    529.06µ ± 0%    31.82µ ±  0%   -93.98% (p=0.000 n=10)
SanitizeHeaders/proto-format_unique_headers-10       4.047m ± 0%   34.068m ± 65%  +741.84% (p=0.000 n=10)
SanitizeHeaders/proto-format_duplicate_headers-10   591.35µ ± 0%    96.74µ ± 33%   -83.64% (p=0.000 n=10)
geomean                                              918.2µ         240.3µ         -73.83%

```

> How about we just don't add it to the generated header [...]

I'm doing something similar here, where instead of entirely removing the multi-length type strings, we append their enum values instead (unit-length). This means we won't need to search for the last " " anymore to trim that off, while still retaining the type data within the header required to convert the header back to the expected representation.

However, either way (removing the type entirely, or appending a known-length suffix) we see that the tests need to be reformatted dramatically (lots of repeated copy-pasting), so I was wondering I'll hold off on doing that until we have an approach that we can go forward with.