@rexagod 2023-12-15
Posting baseline benchmarks, forgot to do this earlier.

* Baseline (same as `main`) v/s changes in this PR.
```
goos: darwin
goarch: arm64
pkg: k8s.io/kube-state-metrics/v2/pkg/metrics_store
                                                  │   BASELINE   │                   HEAD                    │
                                                  │    sec/op    │     sec/op      vs base                   │
SanitizeHeaders/text-format_unique_headers-10       251.62µ ± 0%     31.17µ ±  0%     -87.61% (p=0.000 n=10)
SanitizeHeaders/text-format_duplicate_headers-10    248.52µ ± 4%     31.22µ ±  0%     -87.44% (p=0.000 n=10)
SanitizeHeaders/proto-format_unique_headers-10       250.5µ ± 0%   34093.4µ ± 49%  +13509.75% (p=0.000 n=10)
SanitizeHeaders/proto-format_duplicate_headers-10   247.63µ ± 1%     95.04µ ± 41%     -61.62% (p=0.000 n=10)
geomean                                              249.6µ          237.0µ            -5.05%
```

* `main` + skip empty headers v/s changes in this PR (`main` outperforms this PR when protobuf sanitization is the only thing different in this PR).

```
goos: darwin
goarch: arm64
pkg: k8s.io/kube-state-metrics/v2/pkg/metrics_store
                                                  │   BASELINE   │                   HEAD                    │
                                                  │    sec/op    │     sec/op      vs base                   │
SanitizeHeaders/text-format_unique_headers-10       249.42µ ± 0%     31.14µ ±  0%     -87.51% (p=0.000 n=10)
SanitizeHeaders/text-format_duplicate_headers-10     42.73µ ± 0%     31.12µ ±  0%     -27.16% (p=0.000 n=10)
SanitizeHeaders/proto-format_unique_headers-10       251.1µ ± 0%   36649.4µ ± 64%  +14492.68% (p=0.000 n=10)
SanitizeHeaders/proto-format_duplicate_headers-10    42.74µ ± 0%     93.45µ ± 38%    +118.64% (p=0.000 n=10)
geomean                                              103.4µ          240.0µ          +132.09%
```

To summarize, we optimized the flow with the [first one](https://github.com/kubernetes/kube-state-metrics/pull/2270/files#diff-60450a33adea08c953656dd1e78a80e9f3b279bbc7656dedf31fd1a0c7fc1196R94-R97) quite a bit (to the point where it negated the 1.3x more impact on `SanitizeHeaders`, and gave us a 0.05x boost), but if both `main` and this PR have that in common, with the only feature differentiating this PR being the `proto` sanitization, that impacts the performance compared to the baseline.

Please note that these are the stats compared to the baseline, and not the `regex` implementation (that's present [here](https://github.com/kubernetes/kube-state-metrics/pull/2270#discussion_r1427005713)).