@rexagod 2023-12-15
Added in a couple of O(1) checks from the first commit back in. The benchmarks may differ slightly, but the overall inference should be the same.