@mrueg 2023-12-18
Thanks for running those benchmarks, this is super helpful to get confidence.