@rexagod 2023-12-20
Pasting the requested information below.

<details>
<summary>Testing locally on a Prometheus Protobuf setup.</summary>

<img width="1728" alt="Screenshot 2023-12-20 at 17 05 11" src="https://github.com/kubernetes/kube-state-metrics/assets/33557095/754d0af7-db3b-4495-b6fd-7f476f5cb792">
<img width="1728" alt="Screenshot 2023-12-20 at 17 07 14" src="https://github.com/kubernetes/kube-state-metrics/assets/33557095/60c73bab-6431-4858-a8d1-7fe3aeb3d2d0">

</details>

<details>
<summary>Benchmarking this PR against `main`.</summary>

```
goos: darwin
goarch: arm64
pkg: k8s.io/kube-state-metrics/v2/pkg/metrics_store
                                                  │    BASELINE     │                 HEAD                  │
                                                  │     sec/op      │    sec/op     vs base                 │
SanitizeHeaders/text-format_unique_headers-10           255.0µ ± 3%   1225.8µ ± 3%  +380.70% (p=0.000 n=10)
SanitizeHeaders/text-format_duplicate_headers-10    258262.50n ± 4%    10.87n ± 0%  -100.00% (p=0.000 n=10)
SanitizeHeaders/proto-format_unique_headers-10                         2.071m ± 0%
SanitizeHeaders/proto-format_duplicate_headers-10                      18.50n ± 1%
geomean                                                 256.6µ         4.754µ        -98.58%
```

</details>